fix: use %zu for size_t wx log formatting

+++ src/NexusClient.cpp
@@ -387,9 +387,9 @@ bool NexusClient::ListComponentVersions(const std::string& repositoryBrowseUrl,
         return false;
     }
 
-    wxLogMessage("[nexus] discovered versions component='%s' count=%llu",
+    wxLogMessage("[nexus] discovered versions component='%s' count=%zu",
                  componentName.c_str(),
-                 static_cast<unsigned long long>(outVersions.size()));
+                 outVersions.size());
     // Keep per-component version logging bounded to avoid excessive console spam on large repos.
     constexpr std::size_t kMaxLoggedVersions = 20;
     const auto toLog = std::min(outVersions.size(), kMaxLoggedVersions);
@@ -400,9 +400,9 @@ bool NexusClient::ListComponentVersions(const std::string& repositoryBrowseUrl,
                      version.c_str());
     }
     if (outVersions.size() > kMaxLoggedVersions) {
-        wxLogMessage("[nexus] additional versions omitted component='%s' omittedCount=%llu",
+        wxLogMessage("[nexus] additional versions omitted component='%s' omittedCount=%zu",
                      componentName.c_str(),
-                     static_cast<unsigned long long>(outVersions.size() - kMaxLoggedVersions));
+                     outVersions.size() - kMaxLoggedVersions);
     }
     return true;
 }
@@ -483,11 +483,11 @@ bool NexusClient::DownloadArtifactTree(const std::string& repositoryBrowseUrl,
     }
 
     wxLogMessage(
-        "[nexus] total assets returned (query='%s')=%llu includeFilters=%llu excludeFilters=%llu",
+        "[nexus] total assets returned (query='%s')=%zu includeFilters=%zu excludeFilters=%zu",
                  prefix.c_str(),
-                 static_cast<unsigned long long>(assets.size()),
-                 static_cast<unsigned long long>(regexIncludes.size()),
-                 static_cast<unsigned long long>(regexExcludes.size()));
+                 assets.size(),
+                 regexIncludes.size(),
+                 regexExcludes.size());
 
     std::vector<std::regex> includeRegexes;
     includeRegexes.reserve(regexIncludes.size());
@@ -576,9 +576,9 @@ bool NexusClient::DownloadArtifactTree(const std::string& repositoryBrowseUrl,
         }
     }
 
-    wxLogMessage("[nexus] filtered matches prefix='%s' count=%llu",
+    wxLogMessage("[nexus] filtered matches prefix='%s' count=%zu",
                  prefix.c_str(),
-                 static_cast<unsigned long long>(matches.size()));
+                 matches.size());
 
     if (matches.empty()) {
         errorMessage = "No assets found for path prefix: " + prefix;
@@ -723,8 +723,7 @@ bool NexusClient::ListAssets(const RepoInfo& repo,
             }
         }
 
-        wxLogMessage("[nexus] browse listing discovered files=%llu",
-                     static_cast<unsigned long long>(discovered));
+        wxLogMessage("[nexus] browse listing discovered files=%zu", discovered);
     }
 
     return true;
